Update jsoncpp to 1.7.7 and openssl to 1.0.2j

+++ worker/deps/libsrtp/srtp/test/rtp_decoder.c
@@ -475,7 +475,7 @@ rtp_decoder_handle_pkt(u_char *arg, const struct pcap_pkthdr *hdr,
     return;
   }
   timersub(&hdr->ts, &dcdr->start_tv, &delta);
-  fprintf(stdout, "%02ld:%02ld.%06lu\n", delta.tv_sec/60, delta.tv_sec%60, delta.tv_usec);
+  fprintf(stdout, "%02ld:%02ld.%06ld\n", delta.tv_sec/60, delta.tv_sec%60, (long)delta.tv_usec);
   hexdump(&dcdr->message, pktsize);
 }
